test(cypress): unskip tests and small improvements
