+++ cypress/integration/mobiles-responsiveness.js
@@ -1,3 +1,4 @@
+import { CallTracker } from 'assert'
 import { data } from '../fixtures/mobile-devices.json'
 
 const faker = require('faker')
@@ -14,7 +15,7 @@ describe('Run responsiveness tests on several devices', () => {
   Cypress.config('pageLoadTimeout', 180000) //adding more time for pageLoadTimeout only for this spec
   Cypress.on('uncaught:exception', (err, runnable) => false) // to bypass Module build failed: Error: ENOENT: No such file or directory issue randomly presented
   data.allDevices.forEach((item) => {
-    it.skip(`Create Account on ${item.description}`, () => {
+    it(`Create Account on ${item.description}`, () => {
       cy.viewport(item.width, item.height)
       cy.createAccountPINscreen(randomPIN)
 
@@ -48,17 +49,14 @@ describe('Run responsiveness tests on several devices', () => {
       cy.importAccount(randomPIN, recoverySeed)
       //Validate profile name displayed
       cy.validateChatPageIsLoaded()
-    })
-
-    it.skip(`Chat Features on ${item.description}`, () => {
-      //Setting viewport
-      cy.viewport(item.width, item.height)
-
-      //Validate profile name displayed
-      cy.validateChatPageIsLoaded()
 
       //Go to conversation
       cy.goToConversation('cypress friend', true)
+    })
+
+    it(`Chat Features on ${item.description}`, () => {
+      //Setting viewport
+      cy.viewport(item.width, item.height)
 
       //Validate message and emojis are sent
       cy.chatFeaturesSendMessage(randomMessage)
@@ -66,47 +64,49 @@ describe('Run responsiveness tests on several devices', () => {
 
       //Validate message can be edited
       cy.chatFeaturesEditMessage(randomMessage, randomNumber)
-    })
 
-    it.skip(`Chat - Marketplace - Coming Soon modal content on ${item.description}`, () => {
-      cy.viewport(item.width, item.height)
+      //Adding all remaining chat features tests in same it block, to avoid changing viewports each time a new it block is started
+      //Chat - Marketplace - Coming Soon modal content
+      cy.log(
+        `Chat - Marketplace - Coming Soon modal content on ${item.description}`,
+      )
       cy.get('[data-cy=toolbar-marketplace]').click()
       cy.validateComingSoonModal()
-    })
 
-    it.skip(`Chat - Marketplace - Coming Soon modal button URL on ${item.description}`, () => {
-      cy.viewport(item.width, item.height)
+      //Chat - Marketplace - Coming Soon modal button URL
+      cy.log(
+        `Chat - Marketplace - Coming Soon modal button URL on ${item.description}`,
+      )
       cy.validateURLComingSoonModal()
-    })
 
-    it.skip(`Chat - Marketplace - Coming Soon modal can be dismissed on ${item.description}`, () => {
-      cy.viewport(item.width, item.height)
+      //Chat - Marketplace - Coming Soon modal can be dismissed
+      cy.log(
+        `Chat - Marketplace - Coming Soon modal can be dismissed on ${item.description}`,
+      )
       cy.closeModal('[data-cy=modal-cta]')
-    })
 
-    it.skip(`Chat - Glyph Pack screen is displayed on ${item.description}`, () => {
-      cy.viewport(item.width, item.height)
+      //Chat - Glyph Pack screen is displayed
+      cy.log(`Chat - Glyph Pack screen is displayed on ${item.description}`)
       cy.chatFeaturesSendGlyph()
       cy.goToLastGlyphOnChat().click()
       cy.validateGlyphsModal()
-    })
 
-    it.skip(`Chat - Glyph Pack - Coming Soon modal on ${item.description}`, () => {
-      cy.viewport(item.width, item.height)
+      //Chat - Glyph Pack - Coming Soon modal
+      cy.log(`Chat - Glyph Pack - Coming Soon modal on ${item.description}`)
       cy.contains('View Glyph Pack').click()
       cy.get('[data-cy=modal-cta]').should('be.visible')
       cy.closeModal('[data-cy=modal-cta]')
-    })
 
-    it.skip(`Chat - Glyph Pack screen can be dismissed on ${item.description}`, () => {
-      cy.viewport(item.width, item.height)
+      //Chat - Glyph Pack screen can be dismissed
+      cy.log(`Chat - Glyph Pack screen can be dismissed on ${item.description}`)
       cy.goToLastGlyphOnChat().click()
       cy.get('[data-cy=glyphs-modal]').should('be.visible')
       cy.closeModal('[data-cy=glyphs-modal]')
-    })
 
-    it.skip(`Chat - Glyphs Selection - Coming soon modal on ${item.description}`, () => {
-      cy.viewport(item.width, item.height)
+      //Chat - Glyphs Selection - Coming soon modal
+      cy.log(
+        `Chat - Glyphs Selection - Coming soon modal on ${item.description}`,
+      )
       cy.get('#glyph-toggle').click()
       cy.get('[data-cy=glyphs-marketplace]').click()
       cy.get('[data-cy=modal-cta]').should('be.visible')
