+++ cypress/integration/pin-unlock-validations.js
@@ -3,8 +3,8 @@ const userPassphrase =
   'useful wedding venture reopen forest lawsuit essence hamster kitchen bundle level tower'
 const randomPIN = faker.internet.password(7, false, /[A-Z]/, 'test') // generate random PIN
 
-describe.skip('Unlock pin should be persisted when store pin is enabled', () => {
-  it.skip('Create Account with store pin disabled', () => {
+describe('Unlock pin should be persisted when store pin is enabled', () => {
+  it('Create Account with store pin disabled', () => {
     //Go to URL, add a PIN and make sure that toggle for save pin is disabled
     cy.createAccountPINscreen(randomPIN, false, false)
 
@@ -25,7 +25,7 @@ describe.skip('Unlock pin should be persisted when store pin is enabled', () => 
     })
   })
 
-  it.skip('Create Account with store pin enabled', () => {
+  it('Create Account with store pin enabled', () => {
     //Go to URL, add a PIN and make sure that toggle for save pin is enabled
     cy.createAccountPINscreen(randomPIN, true, false)
